requirements.txt
streamlit==1.36.0
tensorflow==2.18.0
numpy
pillow
opencv-python
ultralytics==8.3.196  
torch>=2.0.0,<2.4.0 

